Change SIOP API
On further reflection, I think this will be more useful.  It's a bit
slower in the exact case, but we shouldn't be looking these up too often
regardless.

We can always switch out the `reverse` for an `into` in the macro
section if we want that back.  We can even do booth at the expensive of
a bit more memory if it truly matters.