Add Deg generator
I remember now that this was what annoyed me last time I looked at the
previous iteration of Raptor.  `Deg[v]` generator uses the `k_prime`
value.  How is it not in the parameters list?

We can see from the tests that the effect is pretty minimal, but it
still exists.. so include it in the required parameters!