Move regex out of module attributes
We only use them once at compile time, but having it named and seperable
improves clarity.  So we just move it into a variable.

It might even get garbage collected this way!  It looks a little weird,
but that's fine.

I just happened to read that this was being deprecated in Elixir 1.19
due to changes in OTP-28.

It should eventually appear at the URI below.

https://hexdocs.pm/elixir/main/compatibility-and-deprecations.html